refactor: split PersonalResult logic to trait for re-use
